# [mypyc] Tokenizer for printf-style format string