--- a/mypyc/irbuild/format_str_tokenizer.py
+++ b/mypyc/irbuild/format_str_tokenizer.py
@@ -3,6 +3,11 @@
 import re
 from typing import List, Tuple
 
+from mypyc.ir.ops import Value, Integer
+from mypyc.ir.rtypes import c_pyssize_t_rprimitive
+from mypyc.irbuild.builder import IRBuilder
+from mypyc.primitives.str_ops import str_build_op
+
 
 def tokenizer_printf_style(format_str: str) -> Tuple[List[str], List[str]]:
     # printf-style String Formatting:
@@ -35,7 +40,24 @@ def tokenizer_printf_style(format_str: str) -> Tuple[List[str], List[str]]:
     return literals, format_op
 
 
-assert tokenizer_printf_style("I'm %s, id years old") == \
-    (["I'm ", ', id years old'], ['%s'])
-assert tokenizer_printf_style("Test: %i%%, Test: %02d, Test: %.2f") == \
-    (['Test: ', '', ', Test: ', ', Test: ', ''], ['%i', '%%', '%02d', '%.2f'])
+def join_formatted_strings(builder: IRBuilder, literals: List[str],
+                           variables: List[Value], line: int) -> Value:
+
+    # The first parameter is the total size of the following PyObject* merged from
+    # two lists alternatively.
+    result_list: List[Value] = [Integer(0, c_pyssize_t_rprimitive)]
+    for a, b in zip(literals, variables):
+        if a:
+            result_list.append(builder.load_str(a))
+        result_list.append(b)
+    # The split_braces() always generates one more element
+    if literals[-1]:
+        result_list.append(builder.load_str(literals[-1]))
+    # Special case for empty string and literal string
+    if len(result_list) == 1:
+        return builder.load_str("")
+    if not variables and len(result_list) == 2:
+        return result_list[1]
+
+    result_list[0] = Integer(len(result_list) - 1, c_pyssize_t_rprimitive)
+    return builder.call_c(str_build_op, result_list, line)