--- a/mypyc/irbuild/format_str_tokenizer.py
+++ b/mypyc/irbuild/format_str_tokenizer.py
@@ -8,55 +8,71 @@ from mypyc.ir.rtypes import c_pyssize_t_rprimitive
 from mypyc.irbuild.builder import IRBuilder
 from mypyc.primitives.str_ops import str_build_op
 
+# printf-style String Formatting:
+# https://docs.python.org/3/library/stdtypes.html#old-string-formatting
+printf_style_pattern = re.compile(r"""
+    (
+    %                                # Start sign
+    (?:\((?P<key>[^)]*)\))?          # Optional: Mapping key
+    (?P<flag>[-+#0 ]+)?              # Optional: Conversion flags
+    (?P<width>\d+|\*)?               # Optional: Minimum field width
+    (?:\.(?P<precision>\d+|\*))?     # Optional: Precision
+    [hlL]?                           # Optional: Length modifier, Ignored
+    (?P<type>[diouxXeEfFgGcrsa])     # Conversion type
+    | %%)
+    """, re.VERBOSE)
+
 
 def tokenizer_printf_style(format_str: str) -> Tuple[List[str], List[str]]:
-    # printf-style String Formatting:
-    # https://docs.python.org/3/library/stdtypes.html#old-string-formatting
-    pattern = re.compile(r"""
-        (
-        %                                # Start sign
-        (?:\((?P<key>[^)]*)\))?          # Optional: Mapping key
-        (?P<flag>[-+#0 ]+)?              # Optional: Conversion flags
-        (?P<width>\d+|\*)?               # Optional: Minimum field width
-        (?:\.(?P<precision>\d+|\*))?     # Optional: Precision
-        [hlL]?                           # Optional: Length modifier, Ignored
-        (?P<type>[diouxXeEfFgGcrsa])     # Conversion type
-        | %%)
-        """, re.VERBOSE)
+    """Tokenize a printf-style format string using regex.
 
+    Return:
+        A list of string literals and a list of conversion operations
+    """
     literals = []
-    format_op = []
+    format_ops = []
     last_end = 0
 
-    for m in re.finditer(pattern, format_str):
+    for m in re.finditer(printf_style_pattern, format_str):
         cur_start = m.start(1)
         format_tmp = m.group(1)
         literals.append(format_str[last_end:cur_start])
-        format_op.append(format_tmp)
+        format_ops.append(format_tmp)
         last_end = cur_start + len(format_tmp)
 
     literals.append(format_str[last_end:])
 
-    return literals, format_op
+    return literals, format_ops
 
 
 def join_formatted_strings(builder: IRBuilder, literals: List[str],
-                           variables: List[Value], line: int) -> Value:
+                           substitutions: List[Value], line: int) -> Value:
+    """Merge the list of literals and the list of substitutions
+    alternatively using 'str_build_op'.
 
-    # The first parameter is the total size of the following PyObject* merged from
-    # two lists alternatively.
+    Args:
+        builder: IRBuilder
+        literals: The literal substrings of the original format string.
+                  After splitting the original format string, the
+                  length of literals should be exactly one more than
+                  substitutions.
+        substitutions: Result Python strings of each conversion
+        line: line number
+    """
+    # The first parameter for str_build_op is the total size of
+    # the following PyObject*
     result_list: List[Value] = [Integer(0, c_pyssize_t_rprimitive)]
-    for a, b in zip(literals, variables):
+    for a, b in zip(literals, substitutions):
         if a:
             result_list.append(builder.load_str(a))
         result_list.append(b)
-    # The split_braces() always generates one more element
     if literals[-1]:
         result_list.append(builder.load_str(literals[-1]))
+
     # Special case for empty string and literal string
     if len(result_list) == 1:
         return builder.load_str("")
-    if not variables and len(result_list) == 2:
+    if not substitutions and len(result_list) == 2:
         return result_list[1]
 
     result_list[0] = Integer(len(result_list) - 1, c_pyssize_t_rprimitive)