### Description

We add a special path for %-style string formatting, which previously is compiled by calling a % binary op. 

The `translate_str_format_percent_sign` serves as an interface from `transform_op_expr` to a specializer for string formatting. A regex tokenizer for %-style format string would parse the original strings into two lists, one is literals and the other is formatting operations.

Currently we directly match the `List[str]` with conversions only for %s and %d cases. In the following PRs, `FormatOp` would be used for better code structure.

`join_formatted_strings`, shared with `str.format()`, merges two lists into one final string using `str_build_op` which is much faster than `str.join`.

## Test Plan

This PR adds several run tests and irbuilding tests. 

The performance on microbenchmarks is: 

master branch:
```
running str_format_percent_operator
..........
interpreted: 0.407026s (avg of 5 iterations; stdev 1.9%)
compiled:    0.309361s (avg of 5 iterations; stdev 2.2%)

compiled is 1.316x faster
```

this PR:
```
running str_format_percent_operator
..........
interpreted: 0.385362s (avg of 5 iterations; stdev 0.75%)
compiled:    0.139198s (avg of 5 iterations; stdev 0.39%)

compiled is 2.768x faster
```